+++ packages/@expressive-code/plugin-frames/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @expressive-code/plugin-frames
 
+## 0.38.1
+
+### Patch Changes
+
+- Updated dependencies [440bb83]
+  - @expressive-code/core@0.38.1
+
 ## 0.38.0
 
 ### Patch Changes