+++ packages/remark-expressive-code/CHANGELOG.md
@@ -1,5 +1,11 @@
 # remark-expressive-code
 
+## 0.38.1
+
+### Patch Changes
+
+- expressive-code@0.38.1
+
 ## 0.38.0
 
 ### Minor Changes